refactor: #8 relocate packages to improve structure
